+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.1.0-alpha.2 (2025-04-08)
+
+Full Changelog: [v0.1.0-alpha.1...v0.1.0-alpha.2](https://github.com/stainless-api/stainless-api-go/compare/v0.1.0-alpha.1...v0.1.0-alpha.2)
+
+### Features
+
+* **api:** update via SDK Studio ([#4](https://github.com/stainless-api/stainless-api-go/issues/4)) ([f06121f](https://github.com/stainless-api/stainless-api-go/commit/f06121f8e6884e77ac754807cf40c1fc0b285cc6))
+* **api:** update via SDK Studio ([#6](https://github.com/stainless-api/stainless-api-go/issues/6)) ([6c7c653](https://github.com/stainless-api/stainless-api-go/commit/6c7c653a883b2fe8871c50c0b149ff7164c28aeb))
+* **client:** support custom http clients ([#8](https://github.com/stainless-api/stainless-api-go/issues/8)) ([4d0d7ce](https://github.com/stainless-api/stainless-api-go/commit/4d0d7cee30fade3e4b16d149f39106e093222024))
+
+
+### Bug Fixes
+
+* **client:** return error on bad custom url instead of panic ([#7](https://github.com/stainless-api/stainless-api-go/issues/7)) ([5995268](https://github.com/stainless-api/stainless-api-go/commit/59952683d6e4f20a137f916e8db9a1d499db7861))
+
 ## 0.1.0-alpha.1 (2025-04-02)
 
 Full Changelog: [v0.0.1-alpha.0...v0.1.0-alpha.1](https://github.com/stainless-api/stainless-api-go/compare/v0.0.1-alpha.0...v0.1.0-alpha.1)